Standardize on *_Api::listUids() method name for api methods that, well, return Uids.
Kronolith_Api::listUids() functionality changed slightly with this commit - it now
no longer returns events that represent exceptions, and only includes recurring events
once within the requested time period. It looks like the syncml code was the only code
currently calling this method, so there should be no issues with this change.

For the other apps, this only represents a method name change.

Bug: 9253