Revert "Fix botched merges."
This reverts commit 9611207f3688474bea6648a109d53ad22195e687.

Conflicts:
	imp/package.xml
	ingo/package.xml

These are not botched merges.  Per discussion on dev@ a while back, we
made the decision to specifically not add any changelog information to
package.xml since it causes a guaranteed merge conflict (another reason
that merging from master -> develop doesn't make sense).

This is why the --nopackage option to horde-components was added.