Right now _getRelatedTags() is the difference between my task list loading in .5 seconds and it taking 4-5 seconds to load, even after the Horde_History cache improvements. It also is getting tags that aren't from the current view because Future tasks aren't accounted for.
I think if we want to keep the related tasks top bar, we need to put that list
together as we go through the list of tasks, where we're already fetching the
tags for each task and displaying them next to each one. That'll avoid a *lot*
of unnecessary queries.