Fix handling of recurring task completion over EAS.
This fixes two issues:

1) When marking a recurring task complete, EAS creates a new
task, marked as a "deadocur" that becomes detached from the
original series. It then either modifies, or replaces the existing
series, moving the due date to the next active recurrence date. We
were saving the "dead" task with the recurrence info, so if that task
was further edited, we would create a duplicate series on the client.

2) If completing an occurence via Horde, the due date was being sent to
EAS incorrectly, leading to the due date being one occurrence or more
behind the actual value.